refactor: move screens into screens folder
