+++ app/pattern-screen/[id].tsx
@@ -1,0 +1,7 @@
+import PatternScreen from "@/screens/PatternScreen";
+
+const PatternViewContainer = () => {
+  return <PatternScreen />;
+};
+
+export default PatternViewContainer;
