+++ components/PatternView.tsx
@@ -4,29 +4,19 @@ import { Measurements } from "@/types";
 import React, { useMemo } from "react";
 import Svg, { G } from "react-native-svg";
 // import { BackBodicePiece, SleevePiece } from '../logic/blousePatterns'; // You would import these too
-import {
-  BackBodicePiece,
-  FrontBodicePiece,
-  SleevePiece,
-} from "@/src/blousePatterns";
+import { Back, Front, SleevePiece } from "@/src/basicBlousePattern";
 import { View } from "react-native";
-import { PatternPieceView } from "./PatternPieceView";
+import { PatternPiece } from "./patterns/PatternPiece";
 
 interface Props {
   measurements: Measurements;
 }
 
-export const BlousePattern: React.FC<Props> = ({ measurements }) => {
+export const PatternView: React.FC<Props> = ({ measurements }) => {
   // useMemo prevents recalculating the pattern on every render
-  const frontPiece = useMemo(
-    () => new FrontBodicePiece(measurements),
-    [measurements]
-  );
+  const frontPiece = useMemo(() => new Front(measurements), [measurements]);
 
-  const backPiece = useMemo(
-    () => new BackBodicePiece(measurements),
-    [measurements]
-  );
+  const backPiece = useMemo(() => new Back(measurements), [measurements]);
 
   const sleevePiece = useMemo(
     () => new SleevePiece(measurements),
@@ -35,21 +25,21 @@ export const BlousePattern: React.FC<Props> = ({ measurements }) => {
 
   console.log("mesurements", measurements);
   return (
-    <View className="border border-black h-full">
+    <View className="h-full">
       <Svg viewBox="-30 0 80 40" width="100%" height="100%">
         {/* Render the Front Piece */}
         <G>
-          <PatternPieceView data={frontPiece.data} color="lightpink" />
+          <PatternPiece data={frontPiece.data} color="lightpink" />
         </G>
 
         {/* Render the Back Piece, offset to the side */}
         <G x={35}>
-          <PatternPieceView data={backPiece.data} color="lightblue" />
+          <PatternPiece data={backPiece.data} color="lightblue" />
         </G>
 
         {/* Render the Back Piece, offset to the side */}
         <G x={10} y={55}>
-          <PatternPieceView data={sleevePiece.data} color="lightgreen" />
+          <PatternPiece data={sleevePiece.data} color="lightgreen" />
         </G>
       </Svg>
     </View>
